Moved things around & repurposed RuleExecutor from Catalyst
